requirements.txt
alembic==1.18.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.77.0
debugpy==1.8.19
decorator==5.2.1
docker==7.1.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.46.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langchain-core==0.3.0
chromadb==0.4.22
overrides>=7.7.0
chroma-hnswlib==0.7.3
bcrypt>=4.0.1
mmh3>=4.0.1
pypika>=0.48.9
typer>=0.9.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
nest-asyncio==1.6.0
numpy>=1.26.0,<2.0.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
optuna==4.6.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.6.1
Pygments==2.19.2
pyparsing==3.3.1
pypdfium2==5.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
selenium==4.39.0
shap==0.47.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
websocket-client==1.9.0
Werkzeug==3.1.5
wsproto==1.3.2
xgboost==3.1.3
zipp==3.23.0
